Updated MySQL nyet for future desired integration test.
It currently only passes in a contrived state.

The part that does not pass is after falling below quota, mysql
does not report the lower storage usage.  This causes the reinstatement
of privileges to fail, unless the operator manually runs an
ANALYZE TABLE or OPTIMIZE TABLE to reclaim the space.  Which command
is proper will require investigation.  This part of the test will be
removed in a future commit, but is documented here to be restored later.

Also made the following changes:

*Added nil return to test app return for non-existent keys.
*Updated cf-mysql-spec NYET to retry until a timeout is reached,
 allowing the quota enforcement job to run on an unknown interval.

Signed-off-by: Ryan Tang <ryantang@pivotallabs.com>